Split tests for Store base class out from the implementation class tests. Refactored tests for FileStore to reflect that change.
